Move CassandraUtils back to lib
