+++ cassandra/lib/CassandraUtils.js
